requirements.txt
SpeechRecognition==3.10.0
pyttsx3==2.90
pyaudio==0.2.13
vosk==0.3.45
psutil==5.9.5
requests==2.31.0
PySide6==6.5.3
qt-material==2.14
python-dotenv==1.0.0
plyer==2.1.0
pywin32==306
winshell==0.6
pvporcupine==2.2.1
pygame==2.5.2
pytest==7.4.3
pytest-benchmark==4.0.0
pytest-cov==4.1.0
pytest-xdist==3.3.1
black==23.11.0
mypy==1.7.0
pylint==3.0.2
